# Week2/Day1-BasicAuth/swaggerConfig.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Airports API",
    "description": "Get all your airport data here https://raw.githubusercontent.com/WhiteHatLearningProducts/airports/master/airportsData.json",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "if you are using NodeJS your dev port might be 3000"
    },
    {
      "url": "https://airports-api.whitehatcoaches.org.uk",
      "description": "this is the production baseURL"
    }
  ],
  "paths": {
    "/airports": {
      "get": {
        "tags": [
          "airports"
        ],
        "summary": "Get an array of all airports objects",
        "description": "Use me to view all 28,000 airports!\n![airport](https://blog.aci.aero/wp-content/uploads/2019/03/shutterstock_745544935-952x635.jpg)\n",
        "responses": {
          "200": {
            "description": "an array of JSON objects that represent each airport",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "icao": {
                        "type": "string"
                      },
                      "iata": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "country": {
                        "type": "string"
                      },
                      "elevation": {
                        "type": "integer"
                      },
                      "lat": {
                        "type": "number",
                        "format": "double"
                      },
                      "lon": {
                        "type": "number",
                        "format": "double"
                      },
                      "tz": {
                        "type": "string"
                      }
                    },
                    "example": {
                      "icao": "00AK",
                      "iata": "",
                      "name": "Lowell Field",
                      "city": "Anchor Point",
                      "state": "Alaska",
                      "country": "US",
                      "elevation": 450,
                      "lat": 59.94919968,
                      "lon": -151.695999146,
                      "tz": "America/Anchorage"
                    }
                  }
                }
              }
            }
          },
          "301": {
            "description": "Moved Permanently"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorised"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "airports"
        ],
        "summary": "Create an airport object",
        "description": "Use me to add an airport!\n![airport](https://blog.aci.aero/wp-content/uploads/2019/03/shutterstock_745544935-952x635.jpg)\n",
        "responses": {
          "201": {
            "description": "the airport has successfully been created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "icao": {
                        "type": "string"
                      },
                      "iata": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "country": {
                        "type": "string"
                      },
                      "elevation": {
                        "type": "integer"
                      },
                      "lat": {
                        "type": "number",
                        "format": "double"
                      },
                      "lon": {
                        "type": "number",
                        "format": "double"
                      },
                      "tz": {
                        "type": "string"
                      }
                    },
                    "example": {
                      "icao": "00AK",
                      "iata": "",
                      "name": "Lowell Field",
                      "city": "Anchor Point",
                      "state": "Alaska",
                      "country": "US",
                      "elevation": 450,
                      "lat": 59.94919968,
                      "lon": -151.695999146,
                      "tz": "America/Anchorage"
                    }
                  }
                }
              }
            }
          },
          "301": {
            "description": "Moved Permanently"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorised"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/airports/{airportICAO}": {
      "get": {
        "tags": [
          "airport"
        ],
        "summary": "Get an airport by the airport ICAO",
        "description": "Use me to get an airport!\n![airport](https://blog.aci.aero/wp-content/uploads/2019/03/shutterstock_745544935-952x635.jpg)\n",
        "parameters": [
          {
            "in": "path",
            "name": "airportICAO",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ICAO of the aiport"
          }
        ],
        "responses": {
          "200": {
            "description": "All Good - returned the airport object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "icao": {
                        "type": "string"
                      },
                      "iata": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "country": {
                        "type": "string"
                      },
                      "elevation": {
                        "type": "integer"
                      },
                      "lat": {
                        "type": "number",
                        "format": "double"
                      },
                      "lon": {
                        "type": "number",
                        "format": "double"
                      },
                      "tz": {
                        "type": "string"
                      }
                    },
                    "example": {
                      "icao": "00AK",
                      "iata": "",
                      "name": "Lowell Field",
                      "city": "Anchor Point",
                      "state": "Alaska",
                      "country": "US",
                      "elevation": 450,
                      "lat": 59.94919968,
                      "lon": -151.695999146,
                      "tz": "America/Anchorage"
                    }
                  }
                }
              }
            }
          },
          "301": {
            "description": "Moved Permanently"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorised"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "airport"
        ],
        "summary": "Update an airport by the airport ICAO",
        "description": "Use me to update an airport!\n![airport](https://blog.aci.aero/wp-content/uploads/2019/03/shutterstock_745544935-952x635.jpg)\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "icao": {
                      "type": "string"
                    },
                    "iata": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "country": {
                      "type": "string"
                    },
                    "elevation": {
                      "type": "integer"
                    },
                    "lat": {
                      "type": "number",
                      "format": "double"
                    },
                    "lon": {
                      "type": "number",
                      "format": "double"
                    },
                    "tz": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "icao": "00AK",
                    "iata": "",
                    "name": "Lowell Field",
                    "city": "Anchor Point",
                    "state": "Alaska",
                    "country": "US",
                    "elevation": 450,
                    "lat": 59.94919968,
                    "lon": -151.695999146,
                    "tz": "America/Anchorage"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "airportICAO",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ICAO of the aiport"
          }
        ],
        "responses": {
          "200": {
            "description": "All Good - Updated the airport object"
          },
          "301": {
            "description": "Moved Permanently"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorised"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "airport"
        ],
        "summary": "Delete an airport by the airport ICAO",
        "description": "Use me to delete an airport!\n![airport](https://blog.aci.aero/wp-content/uploads/2019/03/shutterstock_745544935-952x635.jpg)\n",
        "parameters": [
          {
            "in": "path",
            "name": "airportICAO",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ICAO of the aiport"
          }
        ],
        "responses": {
          "200": {
            "description": "All Good - Deleted the airport object"
          },
          "301": {
            "description": "Moved Permanently"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorised"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}